+++ src/main/kotlin/com/github/morinoparty/morinoinsect/catching/SpawnType.kt
@@ -1,0 +1,5 @@
+package com.github.morinoparty.morinoinsect.catching
+
+enum class SpawnType {
+    ONGROUND, ONTREE
+}
