--- src/main/kotlin/com/github/morinoparty/morinoinsect/catching/area/SpawnType.kt
@@ -1,5 +1,0 @@
-package com.github.morinoparty.morinoinsect.catching.area
-
-enum class SpawnType {
-    ONGROUND, ONTREE
-}
