+++ src/main/kotlin/com/github/morinoparty/morinoinsect/catching/condition/Condition.kt
@@ -1,6 +1,6 @@
 package com.github.morinoparty.morinoinsect.catching.condition
 
-import com.github.morinoparty.morinoinsect.catching.area.SpawnType
+import com.github.morinoparty.morinoinsect.catching.SpawnType
 import org.bukkit.Material
 import org.bukkit.entity.Player
